requirements.txt
fastapi
pydantic
chromadb
sentence-transformers
groq
python-dotenv
uvicorn